Longitudinal Tune equal to stock
I tested this adjustment and it was desirablely equal to the STOCK ACC, in heavy traffic the cars do not cut and nor honk complaining that it does not start to move as soon as the front car begins to move. When you're on the move following a car in front, you don't look "loose" and you don't get jerk acceleration. The distance when the car to the rear of the other is equal to the stock or it is possible to see the rear tires of the car in front. References of master: https://github.com/commaai/openpilot/pull/23441/commits/f1448ce015fc69176acce66505766fc1ee558995