Update Github Actions workflow to split build and publishing steps
This commit splits the existing Github Actions workflow into two distinct workflows: `sunnypilot-build-prebuilt.yaml` and `sunnypilot-publish-prebuilt.yaml`.

In `sunnypilot-build-prebuilt.yaml`, irrelevant sections, `publish-public` and `notify` are removed. The new `branch_name` is added to the `env` section to improve traceability.

The new `sunnypilot-publish-prebuilt.yaml` includes `prepare`, `publish` and `notify` steps. This separation allows to manage the build and release processes independently, and helps improve the maintainability and readability of the workflows.