ui: param watcher
To call this in your panel,
You can use this on offroad panels on _update_state to decrease params access by four times, while remaining responsive without visible lag.

Higher values tested were 0.5, and while that is more efficient, it is also more apparent to the user and makes it seem like the UI is lagging, where in reality, it's just waiting half a second to visually the param they toggled, which is more efficient, but less comfortable ux experience.